+++ src/main/java/com/rcextract/minecord/MinecordPlugin.java
@@ -1,5 +1,8 @@
 package com.rcextract.minecord;
 
+import java.util.Set;
+
+import org.bukkit.event.Listener;
 import org.bukkit.plugin.Plugin;
 
 import com.rcextract.minecord.bukkitminecord.PreferencesManager;
@@ -8,7 +11,7 @@ import com.rcextract.minecord.getters.ServerGetter;
 
 import net.milkbowl.vault.permission.Permission;
 
-public interface MinecordPlugin extends Plugin, ServerGetter, SendableGetter, ConfigurationManager {
+public interface MinecordPlugin extends Plugin, Listener, ServerGetter, SendableGetter, ConfigurationManager {
 
 	public Server getMain();
 	public void setMain(Server server);
@@ -17,8 +20,7 @@ public interface MinecordPlugin extends Plugin, ServerGetter, SendableGetter, Co
 	public DataManipulator getDataManipulator();
 	public void setDataManipulator(DataManipulator dataManipulator);
 	public PreferencesManager getPreferencesManager();
-	public CommandExpansion getCommandExpansion();
-	public void setCommandExpansion(CommandExpansion commandExpansion);
+	public Set<CommandExpansion> getCommandExpansions();
 	public Permission getPermissionManager();
 	public Boolean isUpdateAvailable();
 	public String databaseVersion();