+++ src/main/java/com/rcextract/minecord/RankManager.java
@@ -13,6 +13,7 @@ import com.rcextract.minecord.event.RankCreateEvent;
  * The rank manager helps to manage permission of each user. There is currently no plan on how 
  * it will work, or even it will be removed. It is deprecated temporarily due to no usage.
  */
+@Deprecated
 public class RankManager {
 
 	protected Set<Rank> ranks;