+++ src/main/java/com/rcextract/minecord/Sendable.java
@@ -2,10 +2,9 @@ package com.rcextract.minecord;
 
 import java.util.Set;
 
-import com.rcextract.minecord.getters.ChannelOptionsGetter;
 import com.rcextract.minecord.sql.DatabaseSerializable;
 
-public interface Sendable extends ChannelOptionsGetter, DatabaseSerializable {
+public interface Sendable extends DatabaseSerializable {
 	
 	public int getIdentifier();
 	public String getName();
@@ -15,4 +14,5 @@ public interface Sendable extends ChannelOptionsGetter, DatabaseSerializable {
 	public Set<Server> getServers();
 	public Channel getMain();
 	public void setMain(Channel channel);
+	public void chat(String message);
 }