+++ src/main/java/com/rcextract/minecord/SendableOptions.java
@@ -2,12 +2,18 @@ package com.rcextract.minecord;
 
 import java.util.Arrays;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
 
 import org.apache.commons.lang.Validate;
 import org.bukkit.permissions.Permission;
 
-public class SendableOptions {
+import com.rcextract.minecord.sql.DatabaseSerializable;
+import com.rcextract.minecord.sql.SerializableAs;
+import com.rcextract.minecord.utils.ArrayMap;
+
+@SerializableAs("soptions")
+public class SendableOptions implements DatabaseSerializable {
 
 	private final Sendable sendable;
 	private JoinState state;
@@ -21,6 +27,15 @@ public class SendableOptions {
 		this.state = state;
 		this.rank = rank;
 		this.permissions = new HashSet<Permission>(Arrays.asList(permissions));
+	}
+	
+	@SuppressWarnings("unchecked")
+	public SendableOptions(ArrayMap<String, Object> map) {
+		Map<String, Object> internal = map.toMap();
+		this.sendable = (Sendable) internal.get("sendable");
+		this.state = JoinState.valueOf((String) internal.get("state"));
+		this.rank = (Rank) internal.get("rank");
+		this.permissions = (Set<Permission>) internal.get("permissions");
 	}
 	
 	public Sendable getSendable() {
@@ -46,6 +61,16 @@ public class SendableOptions {
 			if (server.getSendableOptions().contains(this)) 
 				return server;
 		return null;
+	}
+
+	@Override
+	public ArrayMap<String, Object> serialize() {
+		ArrayMap<String, Object> map = new ArrayMap<String, Object>();
+		map.put("sendable", sendable);
+		map.put("state", state.toString());
+		map.put("rank", rank);
+		map.put("permissions", permissions);
+		return map;
 	}
 	
 }