+++ src/main/java/com/rcextract/minecord/event/user/UserMessageEvent.java
@@ -1,57 +1,35 @@
 package com.rcextract.minecord.event.user;
 
 import java.util.Arrays;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
 import org.bukkit.event.HandlerList;
 
 import com.rcextract.minecord.Channel;
-import com.rcextract.minecord.JSONMessage;
+import com.rcextract.minecord.Conversable;
 import com.rcextract.minecord.User;
 
 public class UserMessageEvent extends UserEvent {
 
 	private static final HandlerList handlers = new HandlerList();
-	@Deprecated
-	public static final Set<Integer> REGISTERED_IDENTIFIERS = new HashSet<Integer>();
 	public static HandlerList getHandlerList() {
 		return handlers;
 	}
-	@Deprecated
-	private int id;
 	private String message;
 	private final Channel channel;
-	private final Set<User> recipients;
+	private final Set<Conversable> conversables;
 	private List<UserTagEvent> tags;
-	@Deprecated
-	private JSONMessage finalmessage;
-	@Deprecated
-	public UserMessageEvent(int id, Channel channel, User sender, String message, Set<User> recipients, List<UserTagEvent> tags, JSONMessage finalmessage) {
-		super(sender);
-		this.id = id;
-		REGISTERED_IDENTIFIERS.add(id);
-		this.message = message;
-		this.channel = channel;
-		this.recipients = recipients;
-		this.tags = tags;
-		this.finalmessage = finalmessage;
-	}
-	public UserMessageEvent(String message, Channel channel, User sender, Set<User> recipients, UserTagEvent ... tags) {
+	public UserMessageEvent(String message, Channel channel, User sender, Set<Conversable> conversables, UserTagEvent ... tags) {
 		super(sender);
 		this.message = message;
 		this.channel = channel;
-		this.recipients = recipients;
+		this.conversables = conversables;
 		this.tags = Arrays.asList(tags);
 	}
 	@Override
 	public HandlerList getHandlers() {
 		return handlers;
-	}
-	@Deprecated
-	public int getIdentifier() {
-		return id;
 	}
 	public String getMessage() {
 		return message;
@@ -59,14 +37,10 @@ public class UserMessageEvent extends UserEvent {
 	public Channel getChannel() {
 		return channel;
 	}
-	public Set<User> getRecipients() {
-		return recipients;
+	public Set<Conversable> getConversables() {
+		return conversables;
 	}
 	public UserTagEvent[] getTags() {
 		return tags.toArray(new UserTagEvent[tags.size()]);
-	}
-	@Deprecated
-	public JSONMessage getFinalMessage() {
-		return finalmessage;
 	}
 }